fhir/server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pypdf>=4.2.0
requests>=2.31.0
pydantic>=2.7.0
pyyaml>=6.0.1
python-dateutil>=2.9.0
